Share validation logic between CLI and language server
